Conforms int sizes and removes Y offset.
If we're increasing the offset, why does the rendering seem to move from righ to left?
Here's the explanation I came up with:

What the offset is doing is adding up to the value of the color.
Considering, i.e. on the blue channel, that 0 is black and 255 is blue, it's easy to
observe that the color to be rendererd in any given pixel in the next rendering is a
"bluer" one, and that means, a number closer to 255 by one, compared to the last
iteration, since our offset increases by 1 with every rendering loop.

So, in practice, the first pixel of the screen will be 0, totally black, on the first
rendering; in the second one, the offset will be "1" instead of "0", and it'll be
less black, since it's now "1", and was "0".

As for why the drawing repeats itself, I'll just suppose that, since it's a Uint8,
it's just overflowing when it goes over 255 and resetting to 0.