[Note] Moving on from slow-rendering problem.
I've posted it in StackOverflow and the SDL forums, those didn't help
much. The answers were, mostly:

	- Pixel-manipulation on a texture is costly;
	- Locking/Unlocking is also expensive;
	- I shouldn't re-create the texture with every frame.

For the third problem, I refered to commit 395969ebab77edd7f4409219e4d9f9ba0b59af2f,
explaining that the re-creation was happening because I'd get that segfault
when trying to change its Height above the default size.

I will probably still post it on the Handmade Hero forums, but,
for now, I'll move on with the code.